refactor(front: follow button): フォローボタンのスタイル調整
